Adding tests for StudentT.
