+++ tests/likelihood/gauss_family/test_const_gaussian.py
@@ -1,6 +1,6 @@
 """Unit testsing for ConstGaussian
 """
-import pyccl
+
 import pytest
 import numpy as np
 
@@ -13,36 +13,7 @@ from firecrown.modeling_tools import ModelingTools
 from firecrown.parameters import (
     RequiredParameters,
     DerivedParameterCollection,
-    ParamsMap,
 )
-
-
-@pytest.fixture(name="sacc_with_data_points")
-def fixture_sass_missing_covariance() -> sacc.Sacc:
-    """Return a Sacc object for configuring a ConstGaussian likelihood,
-    but which is missing a covariance matrix."""
-    result = sacc.Sacc()
-    result.add_tracer("misc", "sn_fake_sample")
-    for cnt in [7.0, 4.0]:
-        result.add_data_point("misc", ("sn_fake_sample",), cnt)
-    return result
-
-
-@pytest.fixture(name="sacc_with_covariance")
-def fixture_sacc_with_covariance(sacc_with_data_points: sacc.Sacc) -> sacc.Sacc:
-    result = sacc_with_data_points
-    cov = np.array([[1.0, -0.5], [-0.5, 1.0]])
-    result.add_covariance(cov)
-    return result
-
-
-@pytest.fixture(name="tools_with_vanilla_cosmology")
-def fixture_tools_with_vanilla_cosmology():
-    """Return a ModelingTools object containing the LCDM cosmology from
-    pyccl."""
-    result = ModelingTools()
-    result.update(ParamsMap())
-    result.prepare(pyccl.CosmologyVanillaLCDM())
 
 
 def test_require_nonempty_statistics():
